requirements.txt
fastapi==0.128.0
uvicorn==0.22.0
gspread-formatting==1.2.1
gspread==5.12.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
